requirements.txt
livekit-agents[openai]
livekit-plugins-noise-cancellation
python-dotenv
asyncpg